replace old migration script with new version
