+++ app/src/aws.py
@@ -22,17 +22,6 @@ def md5_for_s3_file(bucket: str, key: str, chunk_size: int = ONE_KB) -> str:
     return md5_hash.hexdigest()
 
 
-def md5_and_size_for_s3_file(bucket: str, key: str, chunk_size: int = ONE_KB) -> tuple[str, int]:
-    response = S3_CLIENT.get_object(Bucket=bucket, Key=key)
-
-    md5_hash = hashlib.md5()
-
-    for chunk in response['Body'].iter_chunks(chunk_size):
-        md5_hash.update(chunk)
-
-    return md5_hash.hexdigest(), response['ContentLength']
-
-
 def list_objects_for_job(bucket: str, job_id: str) -> dict:
     return S3_CLIENT.list_objects_v2(Bucket=bucket, Prefix=job_id)
